Refactoring resizing stage code
